feat (events): Add spider for scraping youngklm events
The meat of the scraping takes place in this class: the `parse` method selects the HTML elements that contain
the events. Then for each event in this list, the child-elements for the url, date, and text are selected. Then the result values of these
elements are returned (yielded asynchronously) to scrapy.